Creating seed file #27

+++ server/db/models/LineItem.js
@@ -10,9 +10,10 @@ const LineItem = db.define('lineitem', {
     allowNull: false
   },
   price: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.INTEGER,
   }
 })
+
 
 LineItem.hook('beforeValidate', (lineItem) => {
   return Product.findOne({where: {id: lineItem.productId }})
@@ -20,7 +21,6 @@ LineItem.hook('beforeValidate', (lineItem) => {
       lineItem.price = theProduct.price * lineItem.quantity
     })
 })
-
 
 
 module.exports = LineItem
